Fix layout of alarm info on the expanded status bar for RTL
The alarm info on the header area in the expanded status bar,
shown after an alarm is set, was incorrectly shown outside
the screen for RTL layouts.

The layout was modified dynamically in Java and the code
didn't consider RTL layouts, always placing the alarm
button to the right of the date.

Change-Id: I3669995c131b47041e5d4ae0f8011da42dfb55ed
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>