Don't block wake locks for required always-on system apps
If an app is marked as "allow-in-power-save" (loaded in as part of
SystemConfig), then this app is not susceptible to battery saving
optimizations and is required to be active for the system to function properly.

If an app is in this list (only com.google.android.gms for now), it will
now not block the wake lock from happening.

Ticket: CYNGNOS-1258

Change-Id: Ia8bf9266489719f101114f334fbf205aee314d15
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: placiano <placiano80@gmail.com>