SysUI: reset mCanDismissKeyguard in setBarState
This patch ensures mCanDismissKeyguard is in the correct state when
setBarState is called, which fixes issues with bouncer showing when
calls come in or an alarm goes off.  This also seems to fix an issue
where the bouncer is not shown when fingerprint authenticiation is
used for the first time on boot.

Change-Id: Ie19fd501b334d4faee14f1ddeb101cbec2515440
TICKET: CYNGNOS-2400 NIGHTLIES-2767
Signed-off-by: placiano <placiano80@gmail.com>