Support for new properties to identify country based ecc #s added by RIL
If the device is in a country where the ecclist list has been
updated with the current country's ECC#s. For example if device
is in INDIA and we are checking for (XYZ, am, 101, XYZ) it will
return true since 101 will be found in ecclist but 101 is not
valid ECC# for Armenia(am). There could be a case where this
logic errors. For example 101 is valid ECC for INDIA and PAK
and we are currently in INDIA and CTS passes (XYZ, pk, 101,
XYZ). We should return true/valid ECC# but current logic will
return false. This scenario fails even today, so there is no
regression.

Issue-id:CYNGNOS-2412

Change-Id: I6e19aed8a84bd57244a40e7f1e06d4625d26ec77
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>