UsbDeviceManager: Remove charging from persisted function composition
CM 12.1 supported the function composition "charging" to indicate no
function compositions, only charging.  CM 13.0 does not support this.
Instead, mtp is used (see commits to fix b/21429947 for an explanation
why this is done).

When a user upgrades from 12.1 to 13.0 and the persisted function
composition is "charging", modifications to the function composition,
such as to add adb or mtp, will timeout and fail.  This introduces a
three second latency on the change.  Removing the charging function
when reading the persisted function composition avoids the delay.

Jira: BACON-4912, OPO-765

Change-Id: Ibbf4f7ce12f97aa046fa214383580ef4d191ce25
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>