SysUI: Don't let rogue themes ruin notifications
If we encounter an exception when inflating a notification's views,
and a theme is applied, we should make a second attempt at inflating
the notification without a theme applied in case the theme cause the
issue.  If the exception still occurs, we treat it like we normally
do and allow the app to be killed for posting a bad notification.

Change-Id: I444cf6c78ee43e2978201880957c53eb08e6966d
TICKET: CYNGNOS-2892
Signed-off-by: placiano <placiano80@gmail.com>