Fix unpinning on device with Navigation Bar
The introduction of the long press back to kill option
messed with the stock behavior of the long press back
button to unpin the locked app. Fortunately, all the job
was already done for the no navigation bar case, so all I
did was to extend the same logic to the case where nav bar
is present.

Change-Id: I2921c653c6e9ce51d49bda31f6e9b64654aeaeaa
Signed-off-by: pafcholini <nadyaivanova14@gmail.com>